Adds a slight API to prevent auto running specs
This allows you to defer the execution so that if you include this in a
page where you want to run the tests via another mechanism (ie a button
click).

Immediately invoke `jasmineRunner.noAutoRun()` after you include this
script in the HTML document. Then when ready execute
`jasmineRunner.run()`. The init method gets run regardless to setup the
CSS. If executed like normal nothing should have changed by this commit.